# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Panacea Socios API",
    "description": "API para gesti\u00f3n de socios, sus preferencias y registro de ventas de Panacea. Permite registrar clientes, capturar sus respuestas a preguntas clave y llevar un historial de compras para futuras promociones y regalos.",
    "contact": {
      "name": "Panacea",
      "email": "info@panacea.com"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "2.0.0"
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new socio with email + password",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login with email + password (OAuth2 password grant)",
        "operationId": "login_local_auth_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_local_auth_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/google": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login with Google ID token",
        "operationId": "login_google_auth_google_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleCredentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/facebook": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login with Facebook access token",
        "operationId": "login_facebook_auth_facebook_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacebookCredentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/apple": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login with Apple identity token",
        "operationId": "login_apple_auth_apple_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleCredentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/socios/{socio_id}/preguntas/": {
      "get": {
        "tags": [
          "Preguntas por Socio"
        ],
        "summary": "Listar preferencias del socio",
        "description": "Retorna todas las preferencias registradas para un socio. Requiere autenticaci\u00f3n.",
        "operationId": "list_preguntas_socios__socio_id__preguntas__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "socio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Socio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PreguntaResponse"
                  },
                  "title": "Response List Preguntas Socios  Socio Id  Preguntas  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Preguntas por Socio"
        ],
        "summary": "Agregar o actualizar preferencia",
        "description": "Crea o actualiza una preferencia (clave \u00fanica por socio). Requiere admin.",
        "operationId": "upsert_pregunta_socios__socio_id__preguntas__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "socio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Socio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreguntaCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreguntaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/socios/{socio_id}/preguntas/{clave}": {
      "get": {
        "tags": [
          "Preguntas por Socio"
        ],
        "summary": "Obtener una preferencia por clave",
        "description": "Retorna la respuesta a una clave espec\u00edfica para un socio. Requiere autenticaci\u00f3n.",
        "operationId": "get_pregunta_socios__socio_id__preguntas__clave__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "socio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Socio Id"
            }
          },
          {
            "name": "clave",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ClavePregunta"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreguntaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Preguntas por Socio"
        ],
        "summary": "Actualizar una preferencia",
        "description": "Actualiza el valor de una preferencia existente. Requiere admin.",
        "operationId": "update_pregunta_socios__socio_id__preguntas__clave__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "socio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Socio Id"
            }
          },
          {
            "name": "clave",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ClavePregunta"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreguntaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreguntaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Preguntas por Socio"
        ],
        "summary": "Eliminar una preferencia",
        "description": "Elimina la respuesta a una clave espec\u00edfica para un socio. Requiere admin.",
        "operationId": "delete_pregunta_socios__socio_id__preguntas__clave__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "socio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Socio Id"
            }
          },
          {
            "name": "clave",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ClavePregunta"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppleCredentials": {
        "properties": {
          "identity_token": {
            "type": "string",
            "title": "Identity Token",
            "examples": [
              "eyJraWQiOiJZdyIs..."
            ]
          }
        },
        "type": "object",
        "required": [
          "identity_token"
        ],
        "title": "AppleCredentials"
      },
      "Body_login_local_auth_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "password"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_local_auth_token_post"
      },
      "ClavePregunta": {
        "type": "string",
        "enum": [
          "que_producto_te_gustaria_que_sumemos",
          "que_parte_de_panacea_es_tu_favorita",
          "de_donde_son"
        ],
        "title": "ClavePregunta"
      },
...
      "FacebookCredentials": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "examples": [
              "EAABsbCS..."
            ]
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "FacebookCredentials"
      },
      "GoogleCredentials": {
        "properties": {
          "id_token": {
            "type": "string",
            "title": "Id Token",
            "examples": [
              "eyJhbGciOiJSUzI1NiIs..."
            ]
          }
        },
        "type": "object",
        "required": [
          "id_token"
        ],
        "title": "GoogleCredentials"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PreguntaCreate": {
        "properties": {
          "clave": {
            "$ref": "#/components/schemas/ClavePregunta",
            "examples": [
              "que_producto_te_gustaria_que_sumemos"
            ]
          },
          "valor": {
            "type": "string",
            "minLength": 1,
            "title": "Valor",
            "examples": [
              "Cervezas artesanales"
            ]
          }
        },
        "type": "object",
        "required": [
          "clave",
          "valor"
        ],
        "title": "PreguntaCreate"
      },
      "PreguntaResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "socio_id": {
            "type": "integer",
            "title": "Socio Id"
          },
          "clave": {
            "$ref": "#/components/schemas/ClavePregunta"
          },
          "valor": {
            "type": "string",
            "title": "Valor"
          }
        },
        "type": "object",
        "required": [
          "id",
          "socio_id",
          "clave",
          "valor"
        ],
        "title": "PreguntaResponse"
      },
      "PreguntaUpdate": {
        "properties": {
          "valor": {
            "type": "string",
            "minLength": 1,
            "title": "Valor"
          }
        },
        "type": "object",
        "required": [
          "valor"
        ],
        "title": "PreguntaUpdate"
      },
      "RegisterRequest": {
        "properties": {
          "nombre_apellido": {
            "type": "string",
            "maxLength": 200,
            "title": "Nombre Apellido",
            "examples": [
              "Juan P\u00e9rez"
            ]
          },
          "email": {
            "type": "string",
            "title": "Email",
            "examples": [
              "juan@example.com"
            ]
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password",
            "examples": [
              "s3cr3tPass!"
            ]
          },
          "dni": {
            "type": "string",
            "maxLength": 20,
            "title": "Dni",
            "examples": [
              "30123456"
            ]
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          },
          "fecha_nacimiento": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Nacimiento"
          }
        },
        "type": "object",
        "required": [
          "nombre_apellido",
          "email",
          "password",
          "dni"
        ],
        "title": "RegisterRequest"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "socio_id": {
            "type": "integer",
            "title": "Socio Id"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "socio_id"
        ],
        "title": "TokenResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        }
      }
    }
  }
}